Update radius and margins as those are going to be used on CA... so CGFloat is default

+++ App/Features/Foundation/UI/UI/Components/Style/Margin.swift
@@ -1,8 +1,8 @@
 public struct Margin {
-  public static let smallest = 4
-  public static let small = 8
-  public static let medium = 16
-  public static let big = 20
-  public static let huge = 24
-  public static let `super` = 32
+  public static let smallest = CGFloat(4)
+  public static let small = CGFloat(8)
+  public static let medium = CGFloat(16)
+  public static let big = CGFloat(20)
+  public static let huge = CGFloat(24)
+  public static let `super` = CGFloat(32)
 }
